+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/form/form.component.ts
@@ -31,7 +31,7 @@ export class CdsFormComponent implements OnInit, OnChanges {
 
   langBot: string;
   fields = new Array() as Field[];;
-  URL_to_form_more_info = "https://gethelp.tiledesk.com/articles/tiledesk-chatbot-forms/";
+  URL_to_form_more_info = "https://guide.tiledesk.com/ai-chatbots-and-automation/tutorials/tiledesk-chatbot-forms";
 
   // modal
   displayMODAL = false;